Memcached only : failover pool support
    Added failover pool support.  Allows sending requests to another pool if
a good server isn't available in the original pool.  This provides an alternate
means for replacing cache servers to ejecting them from the continuum, useful
when cache objects have long lifetimes that are vulnerable to stale data.

    Corrected interpretation of auto_eject_hosts to mean just whether failed
hosts are ejected from the continuum or not, but allow heartbeat and reconnect
handling to be done.

Conflicts:
	src/nc_server.c
	src/nc_server.h